requirements.txt
Flask==3.0.0
opencv-python==4.8.0.76
deepface==0.0.99
numpy==1.26.4
bcrypt==4.0.0
requests==2.31.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
tensorflow>=2.12,<2.18
torch>=2.2,<2.12